Address code review comments
