+++ llzk_backend/src/function.rs
@@ -9,7 +9,6 @@ use crate::gen_context::BlockContextStack;
 use crate::gen_context::GenWithCircomScopeHandling;
 use crate::gen_context::NestedBlockInfo;
 use crate::shared::erase_op;
-use crate::shared::generate_scf_if;
 use crate::shared::get_function_type_attribute;
 use crate::shared::is_bool;
 use crate::shared::is_felt;
@@ -504,6 +503,60 @@ where
             .try_for_each(|(name, result)| self.block_ctx.set_named_value(name, result.into()))?;
         Ok(())
     }
+
+    /// Generate one region for either the then-arm or else-arm of a simple scf.if operation.
+    /// Used by [generate_simple_scf_if].
+    /// The `value_gen` function is called to generate the value to be yielded from the arm.
+    fn generate_simple_scf_if_arm<'ast, F>(
+        &mut self,
+        location: Location<'ctx>,
+        value_gen: F,
+    ) -> Result<(Region<'ctx>, Value<'ctx, 'val>)>
+    where
+        F: FnOnce(&mut Self) -> Result<Value<'ctx, 'val>>,
+    {
+        let region = Region::new();
+        let block = region.append_block(Block::new(&[]));
+        self.block_ctx.push(block);
+        let arm_val = value_gen(self)?;
+        self.block_ctx.pop();
+        block.append_operation(scf::r#yield(&[arm_val], location));
+        Ok((region, arm_val))
+    }
+
+    /// Generate a simple scf.if operation that yields the given `then_value` or `else_value`
+    /// depending on the `condition` value. Unlike [gen_scf_if], this assumes that the
+    /// then and else arms do not modify the current block context and only produce values.
+    pub fn generate_simple_scf_if<'ast, F1, F2>(
+        &mut self,
+        codegen: &LlzkCodegen<'ast, 'ctx>,
+        meta: &Meta,
+        condition: Value<'ctx, 'val>,
+        then_value_gen: F1,
+        else_value_gen: F2,
+    ) -> Result<Operation<'ctx>>
+    where
+        F1: FnOnce(&mut Self) -> Result<Value<'ctx, 'val>>,
+        F2: FnOnce(&mut Self) -> Result<Value<'ctx, 'val>>, {
+        let location = codegen.location_from_meta(meta);
+
+        let (then_region, then_value) = self.generate_simple_scf_if_arm(location, then_value_gen)?;
+        let (else_region, else_value) = self.generate_simple_scf_if_arm(location, else_value_gen)?;
+
+        // Ensure then_value and else_value have the same types
+        assert_eq!(
+            then_value.r#type(), else_value.r#type(),
+            "then and else branches of scf.if must have matching value types"
+        );
+
+        Ok(scf::r#if(
+            condition,
+            &[then_value.r#type()],
+            then_region,
+            else_region,
+            location,
+        ))
+    }
 }
 
 /// The [FunctionContext] directly accesses a single [BlockContextStack] for circom scope handling.
@@ -975,10 +1028,10 @@ where
                     }
                 };
                 let scf_if_op =
-                    generate_scf_if(codegen, function, meta, cond_val, |fc| {
-                        Ok(vec![if_true.gen_llzk_in_function(codegen, fc)?])
+                    function.generate_simple_scf_if(codegen, meta, cond_val, |fc| {
+                        Ok(if_true.gen_llzk_in_function(codegen, fc)?)
                     }, |fc |{
-                        Ok(vec![if_false.gen_llzk_in_function(codegen, fc)?])
+                        Ok(if_false.gen_llzk_in_function(codegen, fc)?)
                     })?;
 
                 function.append_op_unnamed_result(scf_if_op)
